requirements.txt
packaging==25.0
pillow==11.3.0
PyMuPDF==1.26.5
pytesseract==0.3.13
